requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
requests>=2.28.0
PyMuPDF>=1.23.0
google-generativeai>=0.4.0